1st Commit after error occurred to the other repo
In the last repo we mistakenly added the .idea directory
and HelloFX.iml to the git VCS instead of putting them in
the .gitignore file. As a result local and remote branches
collided with several merge conflicts, which we could not
resolve. So, we couldn't help opening this new repo.