DeathOption, playerStats, teamStats, innings added
BattingScoreBoard is also added. It's just the corresponding
class of the table batting_score which tracks the performance
of each batsman in a match.
DeathOption : we can choose a death date if a player or
an umpire dies provided that the status of that player or
umpire has to be "died".
playerStats: playerStats shows the career statistics of a
player. it has been implemented in the player update menu.
teamStats: Similar to playerStat it shows all info about
a team.
innings: keeps track of a particular innings of a match.